requirements.txt
aioredis==1.3.1
amqp==2.6.1
argon2-cffi==19.1.0
asgiref==2.3.0
async-timeout==2.0.1
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
awesome-slugify==1.6.5
billiard==3.6.4.0
celery==4.3.0
certifi==2023.11.17
cffi==1.15.1
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
Collectfast==0.6.2
constantly==15.1.0
cryptography==41.0.7
daphne==2.2.5
defusedxml==0.7.1
Django==2.1.7
django-allauth==0.39.1
django-appconf==1.0.6
django-celery-email==2.0.1
django-compressor==2.2
django-contrib-comments==1.9.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.6
django-haystack==2.8.1
django-markdownx==2.0.28
django-redis==4.10.0
django-taggit==1.1.0
elasticsearch==2.4.1
hiredis==2.3.2
hyperlink==21.0.0
idna==2.8
importlib-metadata==6.7.0
incremental==22.10.0
kombu==4.6.11
Markdown==3.4.4
msgpack==0.6.2
mysqlclient==1.4.2.post1
oauthlib==3.2.2
Pillow==6.2.1
pycparser==2.21
python-slugify==3.0.1
python3-openid==3.1.0
pytz==2023.3.post1
rcssmin==1.0.6
redis==3.2.1
regex==2023.12.25
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.0.12
six==1.16.0
sorl-thumbnail==12.5.0
sqlparse==0.4.4
text-unidecode==1.2
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
Unidecode==0.4.21
urllib3==1.24.3
vine==1.3.0
zipp==3.15.0
zope.interface==6.1
importlib-metadata==4.13.0
pytest==4.3.0
pytest-django==3.4.8
Django-test-plus==1.3.0
